Extract main code into new facade
